remove entitywith id and use id from BaseIdentity (#2)

+++ s2w-domain/src/test/java/com/tradisys/odyssey/apg/s2w/store/leveldb/LevelDBTaskStoreTest.java
@@ -46,7 +46,7 @@ public class LevelDBTaskStoreTest {
 
     @Test
     public void t001_saveAssignment() {
-        int firstTaskId = store.insert(firstTestTask);
+        long firstTaskId = store.insert(firstTestTask);
         store.insert(secondTestTask);
 
         store.saveAssignment(firstTaskId, 1);
